V0.3.0

+++ src/ipcMains.js
@@ -1,14 +1,20 @@
 import { app, ipcMain, dialog } from 'electron'
 import fs                       from 'fs-extra'
 import log                      from 'electron-log'
+import DatesService             from '@/services/datesService'
 import SettingsService          from '@/services/settingsService'
 import MaintenenceService       from '@/services/maintenenceService'
+import PublisherService         from '@/services/publisherService'
+import ServiceGroupService      from '@/services/serviceGroupService'
 
 const isDevelopment = process.env.NODE_ENV !== 'production'
 
 export const enableIPC = () => {
-    const maintenenceService = new MaintenenceService()
-    const settingsService    = new SettingsService()
+    const maintenenceService  = new MaintenenceService()
+    const datesService        = new DatesService()
+    const settingsService     = new SettingsService()
+    const publisherService    = new PublisherService()
+    const serviceGroupService = new ServiceGroupService()
 
     /** Main features ***/
     ipcMain.handle('quit', async() => {
@@ -23,7 +29,77 @@ export const enableIPC = () => {
         return app.getVersion()
     })
 
-    /*** Settings store ***/
+    /*** ServiceGroup store */
+    ipcMain.handle('statsServiceGroups', async () => {
+        return await serviceGroupService.stats()
+    })
+
+    ipcMain.handle('getServiceGroup', async (event, data) => {
+        return await serviceGroupService.findOneById(data.id)
+    })
+
+    ipcMain.handle('getServiceGroups', async (event, args) => {
+        return await serviceGroupService.findAll(args)
+    })
+
+    ipcMain.handle('deleteServiceGroup', async (event, args) => {
+        return await serviceGroupService.delete(args)
+    })
+
+    ipcMain.handle('storeServiceGroup', async (event, args) => {
+        return await serviceGroupService.create(args)
+    })
+
+    ipcMain.handle('updateServiceGroup', async (event, id, data) => {
+        return await serviceGroupService.update(id, data)
+    })
+
+    /*** Publishers store */
+    ipcMain.handle('storePublisher', async (event, data) => {
+        return await publisherService.create(data)
+    })
+    ipcMain.handle('updatePublisher', async (event, id, data) => {
+        return await publisherService.update(id, data)
+    })
+
+    ipcMain.handle('getPublisher', async (event, data) => {
+        return await publisherService.findOneById(data.id)
+    })
+
+    ipcMain.handle('getPublishers', async (event, data) => {
+        return await publisherService.find(data)
+    })
+
+    ipcMain.handle('getNewPublishers', async (event) => {
+        return await publisherService.findNew()
+    })
+
+    ipcMain.handle('deletePublisher', async (event, id) => {
+        return await publisherService.delete(id)
+    })
+
+    ipcMain.handle('statsPublishers', async () => {
+        return await publisherService.stats()
+    })
+
+    ipcMain.handle('getContacts', async () => {
+        return await publisherService.contacts()
+    })
+
+    /**
+     * Dates store
+     **/
+    ipcMain.handle('statsDates', async () => {
+        return await datesService.stats()
+    })
+
+    /**
+     * Settings store
+     **/
+    ipcMain.handle('statsSettings', async () => {
+        return await settingsService.stats()
+    })
+
     ipcMain.handle('getSettings', async () => {
         return await settingsService.find()
     })
@@ -34,6 +110,24 @@ export const enableIPC = () => {
 
     ipcMain.handle('updateSettings', async (event, settingsId, data) => {
         return await settingsService.update(settingsId, data)
+    })
+
+    ipcMain.handle('exportData', async (event, args) => {
+        const exportData = {
+            'date' : new Date(),
+            'type' : args.type,
+            'data' : args.data,
+        }
+
+        let options = {
+            title       : 'Spara export',
+            defaultPath : args.name,
+            buttonLabel : 'Spara',
+        }
+
+        dialog.showSaveDialog(null, options).then(({ filePath }) => {
+            fs.writeFileSync(filePath, JSON.stringify(exportData), 'utf-8')
+        })
     })
 
     ipcMain.on('show-confirmation-dialog', (arg) => {
@@ -60,8 +154,63 @@ export const enableIPC = () => {
         eval(arg.function+'()')
     })
 
+    ipcMain.on('import-publisher', () => {
+        let confirmedImport = true
+        let options         = {
+            title       : 'Importera en förkunnare',
+            buttonLabel : 'Importera',
+            filters     : [
+                { name: 'json', extensions: [ 'json' ] },
+            ],
+            properties: [ 'openFile' ],
+        }
+
+        dialog.showOpenDialog(null, options).then(result => {
+            fs.readFile(result.filePaths[ 0 ], async (err, data) => {
+                if (!err) {
+                    let importData = JSON.parse(data.toString())
+
+                    if(importData.type !== 'publisher'){
+                        confirmedImport     = false
+                        const dialogOptions = {
+                            type      : 'info',
+                            buttons   : [ 'OK' ],
+                            defaultId : 0,
+                            title     : 'Felaktig import-fil',
+                            message   : 'Filen du laddade upp är inte en korrekt import-fil för en förkunnare.',
+                        }
+                        dialog.showMessageBox(null, dialogOptions)
+                    }
+
+                    if(confirmedImport){
+                        const responseOptions = {
+                            type      : 'info',
+                            buttons   : [ 'OK' ],
+                            defaultId : 0,
+                            title     : 'Förkunnaren är importerad',
+                            message   : 'Förkunnaren är importerad!',
+                        }
+
+                        try{
+                            await publisherService.create(importData.data)
+                            dialog.showMessageBox(null, responseOptions)
+                        }catch(err){
+                            log.error(err)
+                            responseOptions.title   = 'Okänt fel'
+                            responseOptions.message = 'Okänt fel'
+                            responseOptions.detail  = 'Det gick inte att importera förkunnaren'
+
+                            dialog.showMessageBox(null, responseOptions)
+                        }
+                    }
+                }
+            })
+        })
+    })
+
     ipcMain.on('recover-backup', () => {
-        const userDataPath = isDevelopment ? './db': app.getPath('userData') + '/db'
+        const userDataPath  = isDevelopment ? './db': app.getPath('userData') + '/db'
+        let confirmedBackup = true
 
         let options = {
             title       : 'Återställ från backup',
@@ -78,34 +227,47 @@ export const enableIPC = () => {
                     let recoveryData = JSON.parse(data.toString())
                     let backupFiles  = recoveryData.backup
 
-                    backupFiles.forEach(file => {
-                        const [ firstKey, ...rest ] = Object.keys(file)
-
-                        fs.writeFile(`${userDataPath}/${firstKey}`, file[ firstKey ], 'utf-8', (err, data) => {
-                            if (err){
-                                log.error(err)
-                            }
-                        })
-                    })
-
-                    const responseOptions = {
-                        type      : 'info',
-                        buttons   : [ 'OK' ],
-                        defaultId : 0,
-                        title     : 'Databasen är återställd',
-                        message   : 'Din databas är återställd!',
-                        detail    : 'Applikationen kommer nu att avslutas och när du startar den nästa gång kommer den nya databasen att läsas in.',
+                    if(recoveryData.type !== 'backup'){
+                        confirmedBackup     = false
+                        const dialogOptions = {
+                            type      : 'info',
+                            buttons   : [ 'OK' ],
+                            defaultId : 0,
+                            title     : 'Felaktig backup-fil',
+                            message   : 'Filen du laddade upp är inte en korrekt backup-fil från Secretary.',
+                            detail    : 'Din databas har inte blvit ändrad. Om du vill återföra en backup måste du välja en korrekt backup-fil.',
+                        }
+                        dialog.showMessageBox(null, dialogOptions)
                     }
 
-                    dialog.showMessageBox(null, responseOptions).then(() => {
-                        app.quit()
-                    })
+                    if(confirmedBackup){
+                        backupFiles.forEach(file => {
+                            const [ firstKey, ...rest ] = Object.keys(file)
+
+                            fs.writeFile(`${userDataPath}/${firstKey}`, file[ firstKey ], 'utf-8', (err, data) => {
+                                if (err){
+                                    log.error(err)
+                                }
+                            })
+                        })
+
+                        const responseOptions = {
+                            type      : 'info',
+                            buttons   : [ 'OK' ],
+                            defaultId : 0,
+                            title     : 'Databasen är återställd',
+                            message   : 'Din databas är återställd!',
+                            detail    : 'Applikationen kommer nu att avslutas och när du startar den nästa gång kommer den nya databasen att läsas in.',
+                        }
+
+                        dialog.showMessageBox(null, responseOptions).then(() => {
+                            app.quit()
+                        })
+                    }
+
                 }
             })
         })
-
-
-
     })
 
     /**
@@ -118,9 +280,12 @@ export const enableIPC = () => {
         const dateString   = date.toLocaleDateString()
         const userDataPath = isDevelopment ? './db': app.getPath('userData') + '/db'
 
+        datesService.upsert('backup', date)
+
         // generate backup file
         const backup = {
             'date'   : date,
+            'type'   : 'backup',
             'backup' : [],
         }
